# Add deferred decoding for array values