--- a/runtime/runtime.go
+++ b/runtime/runtime.go
@@ -2479,7 +2479,7 @@ func NewBlockValue(block Block) interpreter.BlockValue {
 	for i, b := range block.Hash {
 		values[i] = interpreter.UInt8Value(b)
 	}
-	idValue := &interpreter.ArrayValue{Values: values}
+	idValue := interpreter.NewArrayValue(values)
 
 	// timestamp
 	// TODO: verify