Closes #846

## Description

Adds a getter for array elements, that ensures the elements are loaded before accessing them.

This loads all elements of the array, even if one of the elements is accessed. Can be further improved to load only up to that element.

### Performance:
For an ArrayValue with 100 elements, where each element is a struct:
(_Note this includes the performance improvements gained by lazily-loading the composite values as well._)

```
name                                            old time/op    new time/op    delta
ArrayDeferredDecoding/Simply_decode-12             502µs ±20%      41µs ± 0%  -91.80%  (p=0.008 n=5+5)
ArrayDeferredDecoding/Get_first_element-12         514µs ±16%      99µs ± 3%  -80.75%  (p=0.008 n=5+5)
ArrayDeferredDecoding/Get_last_element-12          487µs ± 6%     100µs ± 4%  -79.45%  (p=0.008 n=5+5)
ArrayDeferredDecoding/Re-encode_decoded-12         139µs ±16%       5µs ± 0%  -96.46%  (p=0.008 n=5+5)

name                                            old alloc/op   new alloc/op   delta
ArrayDeferredDecoding/Simply_decode-12             231kB ± 0%       0kB ± 0%  -99.80%  (p=0.008 n=5+5)
ArrayDeferredDecoding/Get_first_element-12         231kB ± 0%      23kB ± 0%  -89.93%  (p=0.008 n=5+5)
ArrayDeferredDecoding/Get_last_element-12          231kB ± 0%      23kB ± 0%  -89.93%  (p=0.008 n=5+5)
ArrayDeferredDecoding/Re-encode_decoded-12        64.6kB ± 0%    33.3kB ± 0%  -48.44%  (p=0.008 n=5+5)

name                                            old allocs/op  new allocs/op  delta
ArrayDeferredDecoding/Simply_decode-12             6.21k ± 0%     0.01k ± 0%  -99.90%  (p=0.008 n=5+5)
ArrayDeferredDecoding/Get_first_element-12         6.21k ± 0%     0.21k ± 0%  -96.59%  (p=0.008 n=5+5)
ArrayDeferredDecoding/Get_last_element-12          6.21k ± 0%     0.21k ± 0%  -96.57%  (p=0.008 n=5+5)
ArrayDeferredDecoding/Re-encode_decoded-12           318 ± 0%        10 ± 0%  -96.86%  (p=0.008 n=5+5)
```

______

For contributor use:

- [x] Targeted PR against `master` branch
- [x] Linked to Github issue with discussion and accepted design OR link to spec that describes this work
- [x] Code follows the [standards mentioned here](https://github.com/onflow/cadence/blob/master/CONTRIBUTING.md#styleguides)
- [x] Updated relevant documentation 
- [x] Re-reviewed `Files changed` in the Github PR explorer
- [x] Added appropriate labels